+++ tests/commands/run_test_utils.py
@@ -30,9 +30,14 @@ class RunAPIMock(requests_mock.Mocker):
         deployment_id=666,
         additional_payload_values=None,
         deployment_version_name="220801.0",
-        num_parameters=None,
+        num_parameters=0,
+        expected_node_count=3,
+        expected_edge_count=5,
     ):
         super().__init__()
+        self.expected_node_count = expected_node_count
+        self.expected_edge_count = expected_edge_count
+        self.num_parameters = num_parameters
         self.last_create_execution_payload = None
         self.last_create_pipeline_payload = None
         self.project_id = project_id
@@ -40,7 +45,6 @@ class RunAPIMock(requests_mock.Mocker):
         self.deployment_id = deployment_id
         self.deployment_version_name = deployment_version_name
         self.additional_payload_values = additional_payload_values or {}
-        self.num_parameters = num_parameters
         self.get(
             f"https://app.valohai.com/api/v0/projects/{project_id}/",
             json=self.handle_project,
@@ -152,8 +156,8 @@ class RunAPIMock(requests_mock.Mocker):
     def handle_create_pipeline(self, request, context):
         body_json = json.loads(request.body.decode("utf-8"))
         assert body_json["project"] == self.project_id
-        assert len(body_json["edges"]) == 5
-        assert len(body_json["nodes"]) == 3
+        assert len(body_json["edges"]) == self.expected_edge_count
+        assert len(body_json["nodes"]) == self.expected_node_count
         if "parameters" in body_json:
             assert len(body_json["parameters"]) == self.num_parameters
         context.status_code = 201